+++ src/core/graph.rs
@@ -1,7 +1,9 @@
+use kdtree::KdTree;
+
 use super::components::options::ComponentsAlgorithmOptions;
+use super::edge::DirectedVehicleSpecificEdgeInformation;
 pub use super::edge::Edge;
 pub use super::node::Node;
-use super::edgeinformation::EdgeInformation;
 use super::routing::RoutingResult;
 use super::routing::options::RoutingAlgorithmOptions;
 pub use super::weight::WeightCalculator;
@@ -10,23 +12,23 @@ use std::collections::HashSet;
 
 use std::rc::Rc;
 
-pub trait Graph {
+pub trait Graph: std::fmt::Debug  {
     //mut functions:
     fn add_node(&mut self, node: Node);
-    fn add_edge(&mut self, edge: Edge);
+    fn add_edge(&mut self, base_node:i32, adj_node: i32,  edge: Edge);
+    fn keep_nodes(&mut self, nodes: &HashSet<i32>);
 
     //simple non mut functions
-    fn get_directed_edge_between(&self, start: i32, end: i32) -> Option<Rc<EdgeInformation>>;
+    fn get_directed_vehicle_specific_edge_information(&self, base_node: i32, adj_node: i32, reverse: bool) ->  Option<Rc<DirectedVehicleSpecificEdgeInformation>>;
     fn get_node(&self,id: i32) -> Option<&Node>;
+
     fn get_nr_nodes(&self) -> usize;
     fn get_nr_edges(&self) -> usize;
-
+    
     //more complex functions
-    fn find_closest_node(&self, lat: f64, lon: f64) -> i32;
-
     fn do_for_all_neighbors<F>(&self, base_node: i32, reverse: bool, f: F)
     where
-        F: FnMut(i32, &Rc<Edge>); //TODO this needs to become independent of actual edge storage
+        F: FnMut(i32);
 
     fn route(&self, opts: &RoutingAlgorithmOptions<Self>, start: i32, end: i32) -> Option<RoutingResult>
     where
@@ -44,7 +46,7 @@ pub trait Graph {
         let mut res = vec![];
         for node in nodes {
             let node = *node;
-            self.do_for_all_neighbors(node, false, |adj_node, _| {
+            self.do_for_all_neighbors(node, false, |adj_node| {
                 //also check if the reverse has been added already
                 //its also possible that there are multiple edges between two nodes so also check the normal order
                 if nodes.contains(&adj_node) && !used.contains(&(adj_node,node)) && !used.contains(&(node, adj_node)) {
@@ -59,4 +61,17 @@ pub trait Graph {
 
         format!("MULTILINESTRING({})",res.join(","))
     } 
+
+    fn create_kd_tree(&self) -> KdTree<f64,i32,[f64;2]>{
+        //last type param are the coordinates, second is the extra data stored and the first is just to clarify the third or something
+        let mut kdtree =  KdTree::new(2);
+
+        for i in 0..self.get_nr_nodes() {
+            let index = i as i32;
+            let node = self.get_node(index).unwrap();
+            kdtree.add([node.lat, node.lon], index).unwrap();
+        }
+        kdtree
+    }
 }
+