+++ src/core/path.rs
@@ -1,13 +1,14 @@
 use std::rc::Rc;
 
-use super::edgeinformation::EdgeInformation;
+use super::{edgeinformation::EdgeInformation, Graph};
 
 //each edge must be linked to the next edge
 pub struct Path {
     edges: Vec<Rc<EdgeInformation>>,
 } 
 
-impl Path {
+impl Path  {
+    //TODO does this need to be Rc's ? 
     pub fn new(edges: Vec<Rc<EdgeInformation>>) -> Self {
         Path::check_edges_valid(&edges, None);
 
@@ -16,9 +17,24 @@ impl Path {
         }
     }
 
+    pub fn add_edge(&mut self, edge: Rc<EdgeInformation>) {
+        let last_node = self.edges.last().map(|e| {
+            e.get_adj_node()
+        });
+
+        if let Some(last_node) = last_node {
+            let base_node = edge.get_base_node();
+            if base_node != last_node {
+                panic!("edges aren't connected: last node {} doesnt match current node {}", last_node, base_node)
+            }
+        }
+
+        self.edges.push(edge);
+    }
+
     pub fn add_edges(&mut self, edges: Vec<Rc<EdgeInformation>>) {
         let last_node = self.edges.last().map(|e| {
-            e.adj_node
+            e.get_adj_node()
         });
 
         Path::check_edges_valid(&edges, last_node);
@@ -35,23 +51,25 @@ impl Path {
                 }
             }
 
-            last_node = Some(edge.adj_node)
+            last_node = Some(edge.get_adj_node())
         }
     }
 
-    pub fn get_wkt(&self) -> String {
+    pub fn get_wkt(&self, graph: &impl Graph) -> String {
         let mut first=false;
         let res = self.edges.iter().fold(vec![],|mut acc, e| {
             if first {
-                let (from_lat, from_lon) = e.get_from_coordinates();
-                acc.push(format!("{:.6} {:.6}",from_lon, from_lat)); //WKT uses lon lat
+                let base_node = e.get_base_node();
+                let node = graph.get_node(base_node).unwrap();
+                acc.push(format!("{:.6} {:.6}",node.lon, node.lat)); //WKT uses lon lat
 
                 first= false;
             }
 
-            let (to_lat, to_lon) = e.get_to_coordinates();
+            let adj_node = e.get_adj_node();
+            let node = graph.get_node(adj_node).unwrap();
 
-            acc.push(format!("{:.6} {:.6}",to_lon, to_lat)); //WKT uses lon lat
+            acc.push(format!("{:.6} {:.6}",node.lon, node.lat)); //WKT uses lon lat
 
             acc
         });
@@ -71,7 +89,7 @@ impl Path {
                 start = false;
             }
 
-            ret.push(edge_info.adj_node);
+            ret.push(edge_info.get_adj_node());
         }
 
         ret