+++ src/core/weight.rs
@@ -1,3 +1,6 @@
+use std::rc::Rc;
+
+use super::edge::DirectedVehicleSpecificEdgeInformation;
 pub use super::edge::Edge;
 
 #[non_exhaustive]
@@ -8,10 +11,10 @@ pub enum WeightCalculator {
 
 impl WeightCalculator {
     #[inline(always)]
-    pub fn calc_weight(&self, edge: &Edge, base_node: i32) -> f64 {
+    pub fn calc_weight(&self, edge: &Rc<DirectedVehicleSpecificEdgeInformation>) -> f64 {
         match self {
-            WeightCalculator::Distance(v) => v.calc_weight(edge, base_node),
-            WeightCalculator::TravelTime(v) => v.calc_weight(edge, base_node),
+            WeightCalculator::Distance(v) => v.calc_weight(edge),
+            WeightCalculator::TravelTime(v) => v.calc_weight(edge),
         }
     }
 }
@@ -20,7 +23,7 @@ pub struct DistanceWeight {}
 
 impl DistanceWeight {
     #[inline(always)]
-    fn calc_weight(&self, edge: &Edge, _base_node: i32) -> f64 {
+    fn calc_weight(&self, edge: &Rc<DirectedVehicleSpecificEdgeInformation>) -> f64 {
         edge.get_distance()
     }
 }
@@ -29,7 +32,7 @@ pub struct TravelTimeWeight {}
 
 impl TravelTimeWeight {
     #[inline(always)]
-    fn calc_weight(&self, edge: &Edge, _base_node: i32) -> f64 {
+    fn calc_weight(&self, edge: &Rc<DirectedVehicleSpecificEdgeInformation>) -> f64 {
         let speed = edge.get_speed();
         if speed > 0.0 {
             return edge.get_distance() / speed;